some architecture refactoring issues
git-svn-id: https://subversion.assembla.com/svn/heartratemonitor@53 bbc9809c-7da1-418a-8ed1-f0bb0a5a736a

+++ trunk/PolarMath/src/org/macondo/polar/evaluation/hrv/IN.java
@@ -3,14 +3,14 @@ package org.macondo.polar.evaluation.hrv;
 import org.macondo.polar.data.Training;
 import org.macondo.polar.evaluation.Evaluation;
 
-public class IN implements Evaluation<Double> {
+public class IN implements Evaluation<Integer> {
 	
-	public Double evaluate(Training training) {
+	public Integer evaluate(Training training) {
 		double bp = new BP().evaluate(training);
-		double amo = new AMoPercents().evaluate(training);
+		int amo = new AMoPercents().evaluate(training);
 		double mo = new Mo().evaluate(training);
 		
-		return amo / (2 * bp * mo);
+		return (int)(amo / (2 * bp * mo));
 	}
 
 }
